+++ tt_metal/impl/dispatch/data_collection.hpp
@@ -84,7 +84,7 @@ void TieRuntimeIdToProgramId(tt_metal::detail::ProgramImpl& program);
 
 // Look up kernel source paths by runtime_id; empty span if the runtime_id is unknown.
 // The returned span is valid until MetalContext teardown or reinitialization.
-std::span<const std::string_view> GetKernelSourcesForRuntimeId(uint16_t runtime_id);
+std::span<const std::string_view> GetKernelSourcesForRuntimeId(uint32_t runtime_id);
 
 // Register a callback to be invoked when real-time profiler data arrives.
 // Multiple callbacks can be registered; they are called in order of registration.
